Restructure and add GLib-based implementations.
This commit restructures the repository into two libraries, libcurvecpr
and libcurvecpr-glib. An out-of-the-box working implementation of the
messager and client are now provided, making use of data structures from
the GLib libraries.

Understandably, some authors may not want to make use of GLib-based
libraries in their projects, so the implementations are provided as a
separate library entire. The --without-glib flag to the configure script
can be used to prevent it from being built.

A basic client implementation using libcurvecpr and libcurvecpr-glib is
now about 150 lines of code, which is actually practical.