Fixed incorrect next timeout calculation.
Previously the next time interval could be set to RTT timeout for
a block (when the block needs to be resent). But when this comes
faster than the write rate computed by Chicago algorithm, such a
premature sendq processing would cause spinning. So we allow a
fast resend only when any acknowledgements are pending as ACKs may
be sent together with blocks faster than the write rate.